youtube-content-automation/requirements.txt
# Core video generation
moviepy==1.0.3
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# Text to Speech - High quality
edge-tts==6.1.9
gtts==2.5.0  # Fallback
pydub==0.25.1

# API clients
requests==2.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0

# Configuration
pyyaml==6.0.1

# Airflow
apache-airflow==2.7.3

# YouTube API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
